requirements-fixed.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
Automat==25.4.16
babel==2.17.0
backports.asyncio.runner==1.2.0
backrefs==5.9
bandit==1.8.6
bcrypt==3.2.0
black==25.1.0
blinker==1.4
boolean.py==5.0
CacheControl==0.14.3
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.0.3
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.9
constantly==15.1.0
coverage==7.10.2
cryptography==45.0.6
cyclonedx-bom==7.0.0
cyclonedx-python-lib==9.1.0
dbus-python==1.2.18
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.0
executing==2.2.0
fancycompleter==0.11.1
fastapi==0.116.1
filelock==3.18.0
flake8==7.3.0
fqdn==1.5.1
ghp-import==2.1.0
h11==0.16.0
httplib2==0.20.2
hyperlink==21.0.0
identify==2.6.12
idna==3.10
importlib-metadata==4.6.4
incremental==24.7.2
iniconfig==2.1.0
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
license-expression==30.4.4
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.16
mkdocs-material-extensions==1.3.1
more-itertools==8.10.0
msgpack==1.1.1
mypy==1.17.1
mypy_extensions==1.1.0
netifaces==0.11.0
nodeenv==1.9.1
oauthlib==3.3.1
packageurl-python==0.17.5
packaging==25.0
paginate==0.5.7
parso==0.8.4
pathspec==0.12.1
pbr==6.1.1
pdbpp==0.11.7
pexpect==4.8.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
py-serializable==2.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyproject-api==1.9.1
pyrepl==0.11.4
pyrsistent==0.18.1
pyserial==3.5
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
pyyaml_env_tag==1.1
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==14.1.0
rpds-py==0.26.0
ruff==0.12.8
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.8.2
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.47.2
stevedore==5.4.1
systemd-python==234
toml==0.10.2
tomli==2.2.1
tox==4.28.4
traitlets==5.14.3
Twisted==25.5.0
types-jsonschema==4.25.0.20250720
types-python-dateutil==2.9.0.20250809
types-PyYAML==6.0.12.20250516
types-setuptools==80.9.0.20250801
typing-inspection==0.4.1
typing_extensions==4.14.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.33.1
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
